Scripts/Telegram_Bot/requirements.txt
detoxify==0.5.0
fastai==2.7.9
fasttext==0.9.2
numpy==1.21.5
requests==2.27.1
python-telegram-bot == 13.13
tensorflow==2.9.1
tensorflow-hub==0.12.0
tensorflow-text==2.9.0
transformers==4.21.1
scipy==1.9.0
scikit-learn==0.22.2.post1